feat: Add game win check
Check whether the player has won by testing if all empty cells
have been clicked on or not. Add the `empty_cells_left` field
to keep track of these cells. In handling each click on empty
 cells through `Msg::Clicked`, `Model`'s update method
decrements the count and checks if it equals zero. If it does
, A `Msg::Win` message is sent. This does introduce a bug,
where the last clicked empty cell will not be revealed, as the
`Msg::Win` is sent before the update completes itself.